+++ test/VaultSyncMode.spec.ts
@@ -9,15 +9,12 @@ import {
   MockUST__factory,
   MockUST,
   MockSyncStrategy,
-  Vault__factory,
 } from '../typechain';
 
 import { depositParams, claimParams } from './shared/factories';
+import createVaultHelpers from './shared/vault';
 
-import {
-  moveForwardTwoWeeks,
-  generateNewAddress,
-} from './shared';
+import { moveForwardTwoWeeks, generateNewAddress } from './shared';
 
 const { parseUnits } = ethers.utils;
 const { MaxUint256 } = ethers.constants;
@@ -27,11 +24,16 @@ describe('Vault in sync mode', () => {
   let alice: SignerWithAddress;
   let bob: SignerWithAddress;
   let carol: SignerWithAddress;
-  let newAccount: SignerWithAddress;
 
   let underlying: MockUST;
   let vault: Vault;
   let strategy: MockSyncStrategy;
+
+  let addUnderlyingBalance: (
+    account: SignerWithAddress,
+    amount: string,
+  ) => Promise<BigNumber>;
+  let addYieldToVault: (amount: string) => Promise<BigNumber>;
 
   const TWO_WEEKS = BigNumber.from(time.duration.weeks(2).toNumber());
 
@@ -43,23 +45,17 @@ describe('Vault in sync mode', () => {
   const fixtures = deployments.createFixture(async ({ deployments }) => {
     await deployments.fixture(['vaults']);
 
-    [owner, alice, bob, carol] = await ethers.getSigners();
+    [owner] = await ethers.getSigners();
 
     const ustDeployment = await deployments.get('UST');
-    const ustVaultDeployment = await deployments.get('Vault_UST');
 
     underlying = MockUST__factory.connect(ustDeployment.address, owner);
-    vault = Vault__factory.connect(ustVaultDeployment.address, owner);
-
-    await addUnderlyingBalance(alice, '1000');
-    await addUnderlyingBalance(bob, '1000');
-    await addUnderlyingBalance(carol, '1000');
   });
 
   beforeEach(() => fixtures());
 
   beforeEach(async () => {
-    [owner, alice, bob, carol, newAccount] = await ethers.getSigners();
+    [owner, alice, bob, carol] = await ethers.getSigners();
 
     let Vault = await ethers.getContractFactory('Vault');
     let MockStrategy = await ethers.getContractFactory('MockSyncStrategy');
@@ -83,6 +79,15 @@ describe('Vault in sync mode', () => {
     strategy = await MockStrategy.deploy(vault.address, underlying.address);
 
     await vault.connect(owner).setStrategy(strategy.address);
+
+    ({ addUnderlyingBalance, addYieldToVault } = createVaultHelpers({
+      vault,
+      underlying,
+    }));
+
+    await addUnderlyingBalance(alice, '1000');
+    await addUnderlyingBalance(bob, '1000');
+    await addUnderlyingBalance(carol, '1000');
   });
 
   describe('#claimYield when there are not enough funds in the vault', () => {
@@ -99,7 +104,7 @@ describe('Vault in sync mode', () => {
     it('claims the yield from the strategy', async () => {
       await addYieldToVault('50');
       await vault.connect(owner).updateInvested();
-      
+
       await vault.connect(alice).claimYield(alice.address);
 
       expect(await underlying.balanceOf(alice.address)).to.eq(
@@ -119,7 +124,7 @@ describe('Vault in sync mode', () => {
       );
     });
 
-    it("rebalances the Vault's reserves when claim > reserves", async () => {
+    it("rebalances the Vault's reserves when yield > reserves", async () => {
       await addYieldToVault('50');
       await vault.connect(owner).updateInvested();
 
@@ -170,7 +175,7 @@ describe('Vault in sync mode', () => {
       await vault.connect(owner).updateInvested();
 
       await moveForwardTwoWeeks();
-      await vault.connect(alice).withdraw(alice.address, [1]);    
+      await vault.connect(alice).withdraw(alice.address, [1]);
 
       expect(await underlying.balanceOf(alice.address)).to.eq(
         '999999999999999999999',
@@ -181,10 +186,10 @@ describe('Vault in sync mode', () => {
       await addYieldToVault('30');
       await vault.connect(owner).updateInvested();
       await addYieldToVault('20');
-  
+
       await moveForwardTwoWeeks();
-      await vault.connect(alice).withdraw(alice.address, [1]);    
-  
+      await vault.connect(alice).withdraw(alice.address, [1]);
+
       expect(await underlying.balanceOf(alice.address)).to.eq(
         '999999999999999999999',
       );
@@ -193,10 +198,10 @@ describe('Vault in sync mode', () => {
     it("rebalances the Vault's reserves when withdraw amount > reserves", async () => {
       await addYieldToVault('50');
       await vault.connect(owner).updateInvested();
-  
+
       await moveForwardTwoWeeks();
-      await vault.connect(alice).withdraw(alice.address, [1]);    
-        
+      await vault.connect(alice).withdraw(alice.address, [1]);
+
       expect(await underlying.balanceOf(vault.address)).to.eq(
         '5000000000000000000',
       );
@@ -216,15 +221,13 @@ describe('Vault in sync mode', () => {
         claims: [claimParams.percent(100).to(bob.address).build()],
       });
       await vault.connect(bob).deposit(bobsDeposit);
-  
+
       await moveForwardTwoWeeks();
-      await vault.connect(alice).withdraw(alice.address, [1]);    
-        
-      expect(await underlying.balanceOf(vault.address)).to.eq(
-        '1',
-      );
+      await vault.connect(alice).withdraw(alice.address, [1]);
+
+      expect(await underlying.balanceOf(vault.address)).to.eq('1');
       expect(await underlying.balanceOf(strategy.address)).to.eq(
-        parseUnits('135')
+        parseUnits('135'),
       );
     });
   });
@@ -245,7 +248,9 @@ describe('Vault in sync mode', () => {
       await vault.connect(owner).updateInvested();
 
       await moveForwardTwoWeeks();
-      await vault.connect(alice).partialWithdraw(alice.address, [1], [parseUnits('50')]);
+      await vault
+        .connect(alice)
+        .partialWithdraw(alice.address, [1], [parseUnits('50')]);
 
       expect(await underlying.balanceOf(alice.address)).to.eq(
         '949999999999999999999',
@@ -255,35 +260,16 @@ describe('Vault in sync mode', () => {
     it("rebalances the Vault's reserves when withdraw amount > reserves", async () => {
       await addYieldToVault('50');
       await vault.connect(owner).updateInvested();
-  
+
       await moveForwardTwoWeeks();
-      await vault.connect(alice).partialWithdraw(alice.address, [1], [parseUnits('50')]);   
-        
-      expect(await underlying.balanceOf(vault.address)).to.eq(
-        parseUnits('10'),
-      );
+      await vault
+        .connect(alice)
+        .partialWithdraw(alice.address, [1], [parseUnits('50')]);
+
+      expect(await underlying.balanceOf(vault.address)).to.eq(parseUnits('10'));
       expect(await underlying.balanceOf(strategy.address)).to.eq(
         '90000000000000000001',
       );
     });
   });
-
-  async function addYieldToVault(amount: string) {
-    await underlying.mint(vault.address, parseUnits(amount));
-    return parseUnits(amount);
-  }
-
-  async function addUnderlyingBalance(
-    account: SignerWithAddress,
-    amount: string,
-  ) {
-    await underlying.mint(account.address, parseUnits(amount));
-    return underlying
-      .connect(account)
-      .approve(vault.address, parseUnits(amount));
-  }
-
-  function removeUnderlyingFromVault(amount: string) {
-    return underlying.burn(vault.address, parseUnits(amount));
-  }
 });
